24조 BE WEEK9 코드리뷰 PR

+++ src/main/java/challenging/application/repository/ParticipantRepository.java
@@ -4,9 +4,7 @@ import challenging.application.domain.Participant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ParticipantRepository extends JpaRepository<Participant, Long> {
-
-  Long countByChallengeId(Long challengeId);
-
+  Integer countByChallengeId(Long challengeId);
   boolean existsByChallengeIdAndMemberId(Long challengeId, Long memberId);
 
 }
